Develop

+++ back/KinoFund/BLL/Collections/CollectionsManager.cs
@@ -1,8 +1,11 @@
-﻿using Core.Dtos.Collections;
+﻿using Core.Dtos.Authentication;
+using Core.Dtos.Collections;
 using Core.Dtos.Movies;
+using Core.Enums;
 using Core.Models;
 using DAL.data;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Query;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -20,20 +23,29 @@ namespace BLL.Collections
             _dbContext = context;
         }
 
-        public async Task<GetAllCollectionsDTO> GetAllAsync()
+        public async Task<GetAllCollectionsDTO> GetAllAsync(string role, long userId)
         {
-            var collections = await _dbContext.Collections
+            IQueryable<CollectionModel> collectionsQuery =  _dbContext.Collections
                 .Include(i => i.Movies)
-                .Include(i => i.User)
-                .Select(c => new CollectionDTO
-                {
-                    CollectionId = c.CollectionId,
-                    Name = c.Name,
-                    Author = c.User.UserName,
-                    Type = c.Type,
-                    MoviesCount = c.Movies.Count
+                .Include(i => i.User);
 
-                }).ToListAsync();
+
+            if(role == AuthConsts.User)
+            {
+                collectionsQuery = collectionsQuery.Where(x => x.Type == CollectionType.Public || x.UserId == userId);
+
+            }
+
+            var collections = await collectionsQuery.Select(c => new CollectionDTO
+            {
+                CollectionId = c.CollectionId,
+                Name = c.Name,
+                Author = c.User.UserName,
+                Type = c.Type,
+                MoviesCount = c.Movies.Count
+
+            }).ToListAsync();
+
 
             return new GetAllCollectionsDTO
             {
@@ -41,7 +53,7 @@ namespace BLL.Collections
             };
         }
 
-        public async Task<CollectionInfoDTO> GetInfoAsync(long collectionId)
+        public async Task<CollectionInfoDTO> GetInfoAsync(long collectionId, string userRole, long userId)
         {
             var collection = await _dbContext.Collections
                 .Include(i => i.Movies)
@@ -50,15 +62,21 @@ namespace BLL.Collections
                 
                 .FirstAsync(c => c.CollectionId == collectionId);
 
+            if(userRole == AuthConsts.User && collection.Type == CollectionType.Private && collection.UserId != userId)
+            {
+                throw new Exception("The collection is private");
+            }
+
             var movies = collection.Movies.Select(x => new CollectionMovieDTO
             { 
                 MovieId = x.MovieId
+
             }).ToList();
 
             return collection.ToDto(movies);
         }
 
-        public async Task<long> CreateAsync(CreateCollectionDTO collectionDTO)
+        public async Task<long> CreateAsync(CreateCollectionDTO collectionDTO, long userId)
         {
             var movieIds = collectionDTO.Movies.Select(x => x.MovieId).ToList();
             var movieModels = await _dbContext.Movies.Where(x => movieIds.Contains(x.MovieId)).ToListAsync();
@@ -68,7 +86,7 @@ namespace BLL.Collections
             {
                 Name = collectionDTO.Name,
                 Type = collectionDTO.Type,
-                UserId = collectionDTO.UserId,
+                UserId = userId,
                 Movies = movieModels
             };
             _dbContext.Collections.Add(collectionModel);
@@ -79,11 +97,16 @@ namespace BLL.Collections
         }
 
 
-        public async Task<bool> EditAsync(EditCollectionDTO collectionDTO)
+        public async Task<bool> EditAsync(EditCollectionDTO collectionDTO, string userRole, long userId)
         {
             var collectionModel = await _dbContext.Collections
                 .Where(c => c.CollectionId == collectionDTO.CollectionId)
                 .FirstOrDefaultAsync();
+
+            if(userRole == AuthConsts.User && userId != collectionModel.UserId)
+            {
+                throw new Exception("You can't modify other people's collections");
+            }
 
             var movieIds = collectionDTO.Movies.Select(x => x.MovieId).ToList();
             var movieModels = await _dbContext.Movies.Where(x => movieIds.Contains(x.MovieId)).ToListAsync();
@@ -102,11 +125,16 @@ namespace BLL.Collections
 
         }
 
-        public async Task<bool> DeleteAsync(long collectionId)
+        public async Task<bool> DeleteAsync(long collectionId, string userRole, long userId)
         {
             var collection = await _dbContext.Collections
                 .Where(c => c.CollectionId == collectionId)
                 .FirstAsync();
+
+            if(userRole == AuthConsts.User && collection.UserId != userId)
+            {
+                throw new Exception("You can't delete other people's collections");
+            }
 
             _dbContext.Collections.Remove(collection);
             await _dbContext.SaveChangesAsync();
